% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{beta_fried1}
\alias{beta_fried1}
\title{(Internal) Computation of the mean value of the regression function in an hyper-rectangle for the \code{Fried1} regression function:}
\usage{
beta_fried1(h)
}
\arguments{
\item{h}{A matrix with two columns (left and right coordinate of each of the hyper-rectangle dimension) and as many rows as the dimension of the number of parameters of the regression function (at least 5 for the \code{Fried1} function), that contains the coordinates of an hyper-rectangle associated to a leaf of a tree}
}
\value{
A vector of length one giving the mean value of the Fried1 regression function in the hyper-rectangle \code{h}.
}
\description{
(Internal) Computation of the mean value of the regression function in an hyper-rectangle for the \code{Fried1} regression function:
}
