% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comput_bias.R
\name{comput_bias}
\alias{comput_bias}
\title{Forest bias computation}
\usage{
comput_bias(
  nbleaves,
  freg.name,
  xdim,
  nbobs = NULL,
  nfor = 1,
  nbobs_test = 500,
  mtry = max(1, floor(xdim/3)),
  tree = FALSE,
  var_estim = FALSE,
  seeds = FALSE,
  mc.cores = 1
)
}
\arguments{
\item{nbleaves}{A vector of the different number of leaves considered to compute the forest bias at}

\item{freg.name}{Name of the regression function to use, among "sum", "fried1", "sinus" and "abs"}

\item{xdim}{Dimension of the input space of th regression function}

\item{nbobs}{The number of observations of the simulated dataset}

\item{nfor}{Number of forests to build (default to 1)}

\item{nbobs_test}{Number of observations of the simulated test set (default to 500)}

\item{mtry}{Number of variables randomly sampled as candidates at each
    split.  Note that the default values are different for
    classification (sqrt(p) where p is number of variables in \code{x})
    and regression (p/3)}

\item{tree}{A logical value: TRUE for a tree bias computation, FALSE for a forest bias computation (defaut to FALSE)}

\item{var_estim}{A logical value to decide to estimate the variance of bias computed for the \code{nfor} forests built}

\item{mc.cores}{Number of cores to give to \code{parallel::mclapply()} function}
}
\value{
A list containing all bias computations. The first component is a matrix summarizing those results with mean and variance estimations for each \code{nbleaves} value. The second argument is a list containing all individual results and the last component is the call of the function.
}
\description{
Forest bias computation
}
\examples{
bias_result <- comput_bias(nbleaves = c(2^5, 2^6), freg.name = "sinus", xdim = 1, nbobs = 640, nbobs_test = 100, nfor = 2, var_estim = TRUE)
}
