% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comput_confint.R
\name{comput_confint}
\alias{comput_confint}
\title{Computation of confidence intervals of biases, and slopes and intercepts estimations by Bootstrap}
\usage{
comput_confint(comput.bias.res, nbBootstrap = 500, alpha = 0.05)
}
\arguments{
\item{comput.bias.res}{Resulting object from \code{comput_bias()} function.}

\item{nbBootstrap}{Number of bootstrap re-samplings used to get the confidence intervals (default to 500).}

\item{alpha}{The probability to not belong to the confidence interval (default to 0.05).}
}
\value{
A list containing the slopes, intercepts and biases confidence intervals.
}
\description{
Computation of confidence intervals of biases, and slopes and intercepts estimations by Bootstrap
}
\examples{
bias_result <- comput_bias(nbleaves = c(2^5, 2^6), freg.name = "sinus", xdim = 1, nbobs = 640, nbobs_test = 100, nfor = 2, var_estim = TRUE)
comput_confint(bias_result, )
}
