% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{descent}
\alias{descent}
\title{Descent of an observation in a tree}
\usage{
descent(z, forest, ind_tree)
}
\arguments{
\item{z}{A vector of the coordinates of the observation we want to let go down the tree}

\item{forest}{The forest structure resulting from the \code{forest_structure} function}

\item{ind_tree}{Index of the tree to consider}
}
\value{
The index of the node where observation \code{z} falls into.
}
\description{
Descent of an observation in a tree
}
\examples{
aforest <- forest_structure(k = 5, xdim = 1, freg.name = "sinus", ntree = 10, nbobs = 100, mtry = 1)
z = 0.5
descent(z, forest = aforest, ind_tree = 1)
}
