% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f_estim}
\alias{f_estim}
\title{Predicts all observations (rows) of a matrix with a forest}
\usage{
f_estim(zmat, forest)
}
\arguments{
\item{zmat}{A matrix with observations to predict in rows}

\item{forest}{The forest structure resulting from the \code{forest_structure} function}
}
\value{
A vector of the predicted values of all rows of the matrix \code{z} by the \code{forest}.
}
\description{
Predicts all observations (rows) of a matrix with a forest
}
\examples{
aforest <- forest_tilde(k = 5, xdim = 1, freg.name = "sinus", ntree = 10, nbobs = 100, mtry = 1)
zmat <- matrix(c(0.2, 0.7), nrow = 2, ncol = 1)
f_estim(zmat, aforest)
}
