% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f_estim_base}
\alias{f_estim_base}
\title{Predicts one observation (one row) of a matrix with a forest}
\usage{
f_estim_base(ind_obs, zmat, forest)
}
\description{
Predicts one observation (one row) of a matrix with a forest
}
