% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{forest_structure}
\alias{forest_structure}
\title{(Internal) Generated a random forest structure build on a simulated dataset using the \code{randomForest::randomForest()} function}
\usage{
forest_structure(
  k,
  xdim,
  freg.name,
  ntree,
  nbobs = NULL,
  mtry = max(1, floor(d/3)),
  seed = NULL
)
}
\arguments{
\item{k}{Number of leaves of the trees (the same for all trees of a forest)}

\item{xdim}{Dimension of the input space of th regression function}

\item{freg.name}{Name of the regression function to use, among "sum", "fried1", "sinus" and "abs"}

\item{ntree}{Number of trees to grow.  This should not be set to too
    small a number, to ensure that every input row gets predicted at
    least a few times. }

\item{nbobs}{The number of observations of the simulated dataset}

\item{mtry}{Number of variables randomly sampled as candidates at each
    split.  Note that the default values are different for
    classification (sqrt(p) where p is number of variables in \code{x})
    and regression (p/3)}

\item{seed}{An optional seed given to the \code{set.seed()} function}
}
\value{
The \code{forest} component of a \code{randomForest} class object from the \code{randomForest} package
}
\description{
(Internal) Generated a random forest structure build on a simulated dataset using the \code{randomForest::randomForest()} function
}
\examples{
aforest <- forest_structure(k = 5, xdim = 1, freg.name = "sinus", ntree = 10, nbobs = 100, mtry = 1)
}
