% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{forest_tilde}
\alias{forest_tilde}
\title{Ideal forest built}
\usage{
forest_tilde(
  k,
  xdim,
  freg.name,
  ntree,
  nbobs = NULL,
  mtry = max(1, floor(xdim/3)),
  seed = NULL
)
}
\arguments{
\item{k}{Number of leaves of the trees (the same for all trees of a forest)}

\item{xdim}{Dimension of the input space of th regression function}

\item{freg.name}{Name of the regression function to use, among "sum", "fried1", "sinus" and "abs"}

\item{ntree}{Number of trees to grow.  This should not be set to too
    small a number, to ensure that every input row gets predicted at
    least a few times. }

\item{nbobs}{The number of observations of the simulated dataset}

\item{mtry}{Number of variables randomly sampled as candidates at each
    split.  Note that the default values are different for
    classification (sqrt(p) where p is number of variables in \code{x})
    and regression (p/3)}

\item{seed}{An optional seed given to the \code{set.seed()} function}
}
\value{
A forest list with the \code{nodepred} component filled with the theoretical values when we know the regression function exactly (hence the name \emph{ideal} forest).
}
\description{
Ideal forest built
}
\examples{
aforest <- forest_tilde(k = 5, xdim = 1, freg.name = "sinus", ntree = 10, nbobs = 100, mtry = 1)
aforest$nodepred
}
