% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{freg}
\alias{freg}
\title{Define the regression function used in the simulation by name}
\usage{
freg(x, freg.name)
}
\arguments{
\item{x}{A vector for one-dimensional functions, a matrix for multidimensional ones, containing values for which the regression function has to be computed}

\item{freg.name}{Name of the regression function to use, among "sum", "fried1", "sinus" and "abs"}
}
\value{
A vector of the resulting computed values of the regression function associated to the inputs in \code{x}.
}
\description{
Define the regression function used in the simulation by name
}
