% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horandomforest-package.R
\docType{package}
\name{horandomforest-package}
\alias{horandomforest}
\alias{horandomforest-package}
\title{horandomforest: Implementation of Hold-Out Random Forests}
\description{
This package contains necessary functions to reproduce the simulation experiments of the article Arlot and Genuer (2026). In this article the rate of convergence of Hold-Out Random Forests are studied. Thus, the package gives an implementation of this Random Forests variant and a plot function to visualize the convergence rate.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robingenuer/horandomforest}
  \item Report bugs at \url{https://github.com/robingenuer/horandomforest/issues}
}

}
\author{
\strong{Maintainer}: Robin Genuer \email{robin.genuer@u-bordeaux.fr}

Authors:
\itemize{
  \item Sylvain Arlot
}

}
\keyword{internal}
