% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hyper_rec}
\alias{hyper_rec}
\title{Hyper-rectangles coordinates computation of the partitions sets of the trees of a forest}
\usage{
hyper_rec(forest, xdim)
}
\arguments{
\item{forest}{The forest structure resulting from the \code{forest_structure} function}

\item{xdim}{Dimension of the input space of th regression function}
}
\value{
An array with same dimensions as \code{hrec} filled with the coordinates of the hyper-rectangles associated to all sets of the partition of the input space (defined by the \code{ind_tree}-th tree of the \code{forest} for \code{hyper_rec_base()} and for all trees of the \code{forest} for \code{hyper_rec()}).
}
\description{
Hyper-rectangles coordinates computation of the partitions sets of the trees of a forest
}
\examples{
aforest <- forest_structure(k = 5, xdim = 1, freg.name = "sinus", ntree = 10, nbobs = 100, mtry = 1)
hyper_rec(forest = aforest, xdim = 1)
}
