% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hyper_rec_base}
\alias{hyper_rec_base}
\title{Hyper-rectangles coordinates computation of the partitions sets of one particular tree of a forest}
\usage{
hyper_rec_base(ind_tree, forest, hrec, xdim)
}
\description{
Hyper-rectangles coordinates computation of the partitions sets of one particular tree of a forest
}
