% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interval_node_bounds}
\alias{interval_node_bounds}
\title{Bounds computation associated to a node of a tree (of a forest)}
\usage{
interval_node_bounds(
  forest,
  ind_tree,
  node,
  dim.int = forest$bestvar[node, ind_tree]
)
}
\arguments{
\item{forest}{The forest structure resulting from the \code{forest_structure} function}

\item{ind_tree}{Index of the tree to consider}

\item{node}{Node of the tree to consider}

\item{dim.int}{a vector of the different dimensions of the input space used for splitting the node, usually (and always for Hold-Out RF) a vector of length one, historically here for dealing for other purely random forests variants}
}
\value{
A vector of length 2 containing the bounds of the interval associated to that node of that tree
}
\description{
Bounds computation associated to a node of a tree (of a forest)
}
\examples{
aforest <- forest_structure(k = 5, xdim = 1, freg.name = "sinus", ntree = 10, nbobs = 100, mtry = 1)
interval_node_bounds(aforest, 1, 3)
}
